% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_daily_files.R
\name{cache_daily_vp_files}
\alias{cache_daily_vp_files}
\title{Cache Daily Vehicle Position Files}
\usage{
cache_daily_vp_files(
  folder = "data",
  folder_save_to = "data",
  timezone = "US/Pacific"
)
}
\arguments{
\item{folder}{A character string specifying the folder path where the RDS files are located. By default, it is set to "data".}

\item{folder_save_to}{A character string specifying the folder path where the condensed file will be saved. By default, it is set to "data".}
}
\value{
This function does not explicitly return a value, but it caches the condensed file(s) in the specified folder.
}
\description{
Combines individual RDS files generated throughout a day into a consolidated file for efficient data storage and retrieval.
}
\details{
The `cache_daily_files` function scans the specified folder for RDS files that contain daily updates, alerts, or variable parameters. It extracts the relevant data from these files and combines them into a single consolidated file. The consolidated file is then saved in the designated folder using a standardized naming convention.

The process involves parsing and filtering the filenames to identify the relevant files, reading each file using the `gauntlet::read_rds_allFiles` function, and merging the data into a unified dataset using the `bind_rows` function from the `dplyr` package. The resulting combined dataset is then saved as an RDS file using the `readr::write_rds` function.

Additionally, the function performs a validation check by reading the saved RDS file using `readr::read_rds` and comparing the number of rows and the content with the original combined dataset. The `check_row` and `check_identical` variables indicate whether the validation is successful.

This function aids in the organization and optimization of data storage, allowing for efficient retrieval and processing of consolidated daily files.
}
\examples{
cache_daily_vp_files(folder = "data", folder_save_to = "cache", timezone = "US/Pacific")

}
\references{
The `gauntlet` and `dplyr` packages are utilized in this function for reading and manipulating data, respectively. Please refer to their documentation for further details.
}
