% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_rtgtfs_query.R
\name{check_data_files}
\alias{check_data_files}
\title{Check if all data files exist and have non-zero size}
\usage{
check_data_files(names, time)
}
\arguments{
\item{names}{a character vector of file names}

\item{time}{a character string representing the sub-query start time in the format "YYYY-MM-DD"}
}
\value{
a list containing two elements: a logical vector indicating if all files exist, and a numeric vector with the size of each file
}
\description{
This function checks if all data files specified by the given \code{names} and \code{sub_query_start_time} exist
and have non-zero size. It returns the result as a list with two elements: a logical vector indicating if all files
exist, and a numeric vector with the size of each file.
}
\details{
This code is not ran by the user and simply checks if the files are made during function \code{query_rtgtfs_json}
}
\examples{
\dontrun{
# Check if files "file1" and "file2" exist for the sub-query starting on January 1st, 2022
check_data_files(c("file1", "file2"), "2022-01-01")
}
}
