% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vp_data_to_sf.R
\name{convert_vp_data_to_sf}
\alias{convert_vp_data_to_sf}
\title{Convert VP Data to Spatial Format}
\usage{
convert_vp_data_to_sf(data, crs_to)
}
\arguments{
\item{data}{A data.frame containing longitude and latitude columns.}

\item{crs_to}{A character string specifying the Coordinate Reference System (CRS) for the specific area where the data is located. The CRS should have meters as its base unit.}
}
\value{
A spatial object of class 'sf' representing the converted data in the specified CRS.
}
\description{
Converts a data.frame containing longitude and latitude coordinates into a spatial format suitable for mapping purposes.
}
\details{
The `convert_vp_data_to_sf_1` function takes a data.frame with longitude and latitude columns and converts it into a spatial object using the 'sf' package. The resulting spatial object can be used for mapping and spatial analysis purposes.

The function first converts the data.frame into an 'sf' object using the `sf::st_as_sf` function, specifying the longitude and latitude columns as the coordinate values and the CRS of the original data as EPSG code 4326 (WGS84). The `crs_to` parameter allows the user to specify the desired CRS for the converted data. It is important to choose a CRS appropriate for the specific area where the data is located and ensure that the CRS has meters as its base unit.

The converted spatial object is then transformed to the specified CRS using the `sf::st_transform` function. The function also applies the `gauntlet::st_extract_coords` function to extract the route, vehicle, and trip identifiers from the data and groups the data by these identifiers.

Additional computations are performed within each group using the `dplyr` package. These computations include calculating the differences in longitude and latitude between consecutive points, the total distance traveled (`ttl_diff`), and the average speed (`speed_avg`). Points with an average speed exceeding 90 meters per second are filtered out to remove any potential outliers. The data is then ungrouped and transformed back to EPSG code 4326 to ensure compatibility with common mapping tools.

This function simplifies the process of converting tabular data with coordinates into a spatial format, enabling easier visualization and analysis of geographic data.
}
\examples{
# Convert VP data in a data.frame format to an 'sf' object with a specific CRS.
converted_data <- convert_vp_data_to_sf(vp_data, crs_to = "EPSG:32633")

}
\references{
The `sf` and `dplyr` packages are used in this function for spatial data manipulation and analysis, respectively. Please refer to their documentation for further details.
}
