% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_shape_xwalk.R
\name{get_gtfs_shape_xwalk}
\alias{get_gtfs_shape_xwalk}
\title{Create a crosswalk dataframe between the realtime GTFS vehicle position dataset and the static GTFS input}
\usage{
get_gtfs_shape_xwalk(vp_data, gtfs)
}
\arguments{
\item{vp_data}{A dataframe containing the realtime GTFS vehicle position dataset.}

\item{gtfs}{A list or dataframe representing the static GTFS input.}
}
\value{
A dataframe representing the crosswalk between trip IDs and shape IDs.
}
\description{
This function takes a pre-processed, realtime GTFS vehicle position dataset and a static GTFS list as inputs. It creates
a crosswalk dataframe that links the trip IDs from the vehicle poisition dataset to their corresponding shape IDs
from the static GTFS input. The function notifies the user about the start and completion of the operation,
and also reports the number of unique shapes and total number of trip IDs found in the crosswalk dataframe.
The crosswalk dataframe is required by the function called `make_speed_profiles()`.
}
\examples{
vp_data <- read.csv("vehicle_position_data.csv")
gtfs <- read_gtfs("static_gtfs.zip")
get_gtfs_shape_xwalk(vp_data, gtfs)

}
