% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_speed_profiles.R
\name{make_speed_profiles}
\alias{make_speed_profiles}
\title{Create speed profiles by snapping realtime GTFS vehicle position data to discretized segments of the static GTFS object}
\usage{
make_speed_profiles(
  sf_object,
  crs,
  xwalk_ob,
  vp_data,
  samp_dist = 100,
  over_ride = NA
)
}
\arguments{
\item{sf_object}{An `sf` object representing the preprocessed static GTFS object.}

\item{crs}{The coordinate reference system (CRS) to be used for snapping.}

\item{xwalk_ob}{The crosswalk object linking trip IDs to shape IDs.}

\item{vp_data}{Realtime GTFS vehicle position data.}

\item{samp_dist}{The sampling distance for snapping the points to discretized segments (default is 100).}

\item{over_ride}{Optional parameter for overriding the snapping process (default is NA).}
}
\value{
A list containing the route samples and speed profiles.
}
\description{
This function takes a preprocessed static GTFS object represented as an `sf` object, a coordinate reference system (CRS),
a crosswalk object linking trip IDs to shape IDs, realtime GTFS vehicle position data, and optional parameters for
sampling distance and overriding the snapping process. The function notifies the user about the start and completion
of the operation and also provides progress messages during the execution of the `map_df` function. It returns a list
containing the route samples and speed profiles.
}
\examples{
sf_object <- read_sf("static_gtfs.shp")
crs <- "+proj=utm +zone=10 +ellps=WGS84"
xwalk_ob <- read_csv("crosswalk.csv")
vp_data <- read_csv("vehicle_position_data.csv")
make_speed_profiles(sf_object, crs, xwalk_ob, vp_data, samp_dist = 100, over_ride = NA)

}
