% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_query_records.R
\name{plot_batch_query_records}
\alias{plot_batch_query_records}
\title{Plot Batch Query Records}
\usage{
plot_batch_query_records(data, days = NA)
}
\arguments{
\item{data}{A data frame containing the batch query records.}

\item{days}{An optional parameter specifying the number of days to include in the plot. Defaults to NA, which includes all available data.}
}
\value{
A ggplot object displaying the batch query records.
}
\description{
This function generates a plot showing the batch query records over time.
}
\examples{
data <- read.csv("batch_query_data.csv")
plot_batch_query_records(data, days = 7)
}
