% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_buses_in_operation.R
\name{plot_buses_in_operation}
\alias{plot_buses_in_operation}
\title{Plot Buses in Operation}
\usage{
plot_buses_in_operation(data, days = 7)
}
\arguments{
\item{data}{A preprocessed data.frame containing vehicle position data.}

\item{days}{An optional numeric value specifying the number of days to consider for plotting. By default, it is set to 7.}
}
\description{
Creates a line plot to visualize the number of buses in operation over time.
}
\details{
The `plot_buses_in_operation` function takes preprocessed vehicle position data and creates a line plot to monitor the number of buses in operation over time. The function aggregates the data by minute, counts the number of buses in operation within each minute, and creates a line plot with time on the x-axis and the number of buses on the y-axis.

By default, the function considers the last 7 days of data. However, the `days` parameter allows the user to specify a different number of days to include in the plot.

The resulting plot provides an overview of the bus operations, allowing for monitoring and analysis of bus activity patterns.
}
\examples{
# Plot the number of buses in operation over the last 7 days
plot_buses_in_operation(vehicle_data)

}
