% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_batch_data.R
\name{plot_query_batch_data}
\alias{plot_query_batch_data}
\title{Plot Query Batch Data}
\usage{
plot_query_batch_data(data, days = 7)
}
\arguments{
\item{data}{A preprocessed data.frame containing vehicle position data.}

\item{days}{An optional numeric value specifying the number of days to consider for plotting. By default, it is set to 7.}
}
\description{
Creates a tile plot to visualize the count of records in query batches over time.
}
\details{
The `plot_query_batch_data` function takes preprocessed vehicle position data and creates a tile plot to monitor the count of records in query batches over time. The function aggregates the data by minute, counts the records within each minute, and creates a tile plot with time on the x-axis and date on the y-axis. Each tile represents a minute and is colored based on the count of records.

By default, the function considers the last 7 days of data. However, the `days` parameter allows the user to specify a different number of days to include in the plot.

The resulting plot provides an overview of the distribution and frequency of records in query batches, enabling the monitoring of data collection and processing processes.
}
\examples{
# Plot the count of records in query batches over the last 7 days
plot_query_batch_data(vehicle_data)

}
