% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_and_process_static_gtfs.R
\name{query_and_process_static_gtfs}
\alias{query_and_process_static_gtfs}
\title{Query and process GTFS data from a user-supplied URL}
\usage{
query_and_process_static_gtfs(url_gtfs)
}
\arguments{
\item{url_gtfs}{The URL of the static GTFS data supplied by the municipal transit operator.}
}
\value{
A spatial data frame representing the processed GTFS data.
}
\description{
This function takes a user-supplied static GTFS URL, provided by municipal transit operators,
as input. It attempts to read and process the GTFS data using the `tidytransit` package. It
notifies the user about the outcome of the operation, including whether it was successful
or if an error occurred during the reading process.
}
\examples{
url_gtfs <- "https://example.com/gtfs.zip"
query_and_process_gtfs(url_gtfs)

}
