% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_rtgtfs_query.R
\name{query_rtgtfs_json}
\alias{query_rtgtfs_json}
\title{Query RTGTFS JSON data}
\usage{
query_rtgtfs_json(
  rtgtfs_feeds,
  ttl_query_duration,
  cache_interval,
  query_interval,
  routes,
  log_file
)
}
\arguments{
\item{rtgtfs_feeds}{A list of RTGTFS feed URLs (containing URLs for vehicle positions, updates, and alerts).}

\item{ttl_query_duration}{The duration of the entire query process (in hours).}

\item{cache_interval}{The interval between subsequent queries for the same data (in minutes).}

\item{query_interval}{The interval between queries for new data (in seconds).}

\item{routes}{A character vector containing the route IDs to be queried.}

\item{log_file}{A character string representing the path to the log file.}
}
\value{
Nothing is returned by this function, but the obtained data is saved as RDS files.
}
\description{
This function queries real-time data for specified routes from RTGTFS feeds, and saves the obtained data as RDS files.
}
\examples{
\dontrun{
query_rtgtfs_json(rtgtfs_feeds = list(url_vp = "http://url_vp", url_updates = "http://url_updates", url_alerts = "http://url_alerts"),
                  ttl_query_duration = 12, cache_interval = 5, query_interval = 10,
                  routes = c("route_1", "route_2"), log_file = "query.log")
}
}
